Minor refactoring and changes to API for documentability
