+++ loomengine/master/api/test/runs/test_run_execution.py
@@ -6,7 +6,7 @@ from api.serializers import FileDataObjectSerializer, \
     RunRequestSerializer, TemplateSerializer
 
 import loomengine.utils.md5calc
-import loomengine.utils.helper
+
 
 class AbstractRunTest(object):
 
@@ -54,9 +54,9 @@ class AbstractRunTest(object):
         s.is_valid(raise_exception=True)
         with self.settings(WORKER_TYPE='MOCK'):
             run_request = s.save()
-        loomengine.utils.helper.wait_for_true(
+        wait_for_true(
             lambda: RunRequest.objects.get(id=run_request.id).run.postprocessing_status == 'complete', timeout_seconds=120, sleep_interval=1)
-        loomengine.utils.helper.wait_for_true(
+        wait_for_true(
             lambda: all([step.postprocessing_status=='complete'
                          for step
                          in RunRequest.objects.get(
@@ -112,3 +112,16 @@ class TestManySteps(TransactionTestCase, AbstractRunTest):
 
     def testRun(self):
         pass
+
+def wait_for_true(test_method, timeout_seconds=20, sleep_interval=None):
+    from datetime import datetime
+    import time
+
+    if sleep_interval is None:
+        sleep_interval = timeout_seconds/10.0
+    start_time = datetime.now()
+    while not test_method():
+        time.sleep(sleep_interval)
+        time_running = datetime.now() - start_time
+        if time_running.seconds > timeout_seconds:
+            raise Exception("Timeout")
